+++ gluecodium/src/main/java/com/here/gluecodium/Gluecodium.kt
@@ -41,6 +41,7 @@ import com.here.gluecodium.validator.LimeFieldConstructorsValidator
 import com.here.gluecodium.validator.LimeFunctionsValidator
 import com.here.gluecodium.validator.LimeGenericTypesValidator
 import com.here.gluecodium.validator.LimeInheritanceValidator
+import com.here.gluecodium.validator.LimeLambdaValidator
 import com.here.gluecodium.validator.LimeOptimizedListsValidator
 import com.here.gluecodium.validator.LimePropertiesValidator
 import com.here.gluecodium.validator.LimeSerializableStructsValidator
@@ -195,6 +196,7 @@ class Gluecodium(
             { LimePropertiesValidator(limeLogger).validate(it) },
             { LimeSkipValidator(limeLogger).validate(it) },
             { LimeAsyncValidator(limeLogger).validate(it) },
+            { LimeLambdaValidator(limeLogger, generatorOptions).validate(it) },
         )
 
     companion object {