+++ CHANGELOG.md
@@ -2,10 +2,12 @@
 
 ## Unreleased:
 ### Features:
+ * Implemented validation of comments used for lambdas. When the description of parameters or return value is missing, then warning is generated. The user may also treat the warning as error via 'werror' flag.
  * Implemented validation of comments used for functions. When the description of parameters or return value is missing, then warning is generated. The user may also treat the warning as error via 'werror' flag.
 ### Bug fixes:
  * Gluecodium.kt: fixed a bug related to running 'LimeFunctionsValidator' twice. It was not needed - running the validator once is sufficient.
- * Dart: fixed a bug related to missing/superflous 'const' keyword usage in definition of default values in constructors that used collections when `@Immutable` or `PositionalDefaults` were specified.
+ * Dart: fixed a bug related to missing 'const' keyword usage in definition of default values in constructors that used immutable structure types when `@Immutable` or `@PositionalDefaults` were specified.
+ * Dart: fixed a bug related to missing/superflous 'const' keyword usage in definition of default values in constructors that used collections when `@Immutable` or `@PositionalDefaults` were specified.
 
 ## 13.9.7
 Release date 2024-11-13